+++ app/not-found.js
@@ -1,0 +1,11 @@
+export const metadata = {
+  title: "Not Found",
+};
+
+export default function NotFound() {
+  return (
+    <div>
+      <h1>What are you doing here...?</h1>
+    </div>
+  );
+}
