+++ package.json
@@ -8,9 +8,10 @@
     "post-update": "yarn upgrade --latest"
   },
   "dependencies": {
-    "next": "^12.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "eslint-config-next": "^13.5.4",
+    "next": "^13.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "license": "MIT",
   "keywords": [],
