--- pages/404.js
@@ -1,14 +1,0 @@
-import Head from "next/head";
-
-export default function NotFound() {
-  return (
-    <>
-      <Head>
-        <title>Not Found | The New York Times Best Seller</title>
-      </Head>
-      <main className="paper">
-        <h1>Not Found...</h1>
-      </main>
-    </>
-  );
-}
