--- pages/_app.js
@@ -1,9 +1,0 @@
-import Layout from "../components/Layout";
-
-export default function App({ Component, pageProps }) {
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  );
-}
